Change ticket price field type from integer to decimal for precision

+++ database/migrations/2026_04_15_091739_create_tickets_table.php
@@ -16,8 +16,7 @@ return new class extends Migration
 
             $table->string('ticket_number')->unique();
             $table->string('rank');
-            $table->integer('price');
-
+            $table->decimal('price', 10, 2);
             $table->timestamps();
 
             // foreign keys
